replace to separated package Proxy* classes
