+++ src/main/java/executor/service/service/proxy/ProxyValidator.java
@@ -1,4 +1,4 @@
-package executor.service.service;
+package executor.service.service.proxy;
 
 import executor.service.model.ProxyConfigHolder;
