+++ src/main/java/executor/service/service/proxy/impl/ProxyValidatorImpl.java
@@ -1,8 +1,8 @@
-package executor.service.service.impl;
+package executor.service.service.proxy.impl;
 
 import executor.service.config.properties.PropertiesConstants;
 import executor.service.model.ProxyConfigHolder;
-import executor.service.service.ProxyValidator;
+import executor.service.service.proxy.ProxyValidator;
 import org.apache.http.HttpHost;
 import org.apache.http.HttpStatus;
 import org.apache.http.auth.AuthScope;
